W-13145685: Modularize `mule-ee` modules
* core-ee and tracking
